Release v9.2.0: Network optimizations, diagnostics, and Keep-Alive hangs
