+++ custom_components/climate_ip/manifest.json
@@ -11,7 +11,9 @@
   "issue_tracker": "https://github.com/ismaelrivas/samsungrac/issues",
   "requirements": [
     "requests>=2.21.0",
-    "xmltodict>=0.13.0"
+    "xmltodict>=0.13.0",
+    "icmplib>=3.0"
   ],
-  "version": "9.1.1"
+  "version": "9.2.0",
+  "homeassistant": "2024.1.0"
 }
